Move SubredditTool to own source file
